Split out all the patches into their own file
