+++ src/patches/removeTracking.ts
@@ -1,0 +1,42 @@
+import { applyPatch, Patch } from './helpers';
+
+const patch: Patch = {
+    id: 'remove-tracking',
+    match: (url: string) => url === 'index.html',
+    async apply(src: string) {
+        // Remove tracking pixels
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*xsolla[^>]*<\/script>/g,
+            '<script>xnt = new Proxy(() => xnt, { get: () => xnt })</script>',
+        );
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*facebook[^>]*<\/script>/g,
+            '<script>fbq = new Proxy(() => fbq, { get: () => fbq })</script>',
+        );
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*google[^>]*<\/script>/g,
+            '<script>ga = new Proxy(() => ga, { get: () => ga })</script>',
+        );
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*mxpnl[^>]*<\/script>/g,
+            '<script>mixpanel = new Proxy(() => mixpanel, { get: () => mixpanel })</script>',
+        );
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*twttr[^>]*<\/script>/g,
+            '<script>twttr = new Proxy(() => twttr, { get: () => twttr })</script>',
+        );
+        src = applyPatch(
+            src,
+            /<script[^>]*>[^>]*onRecaptchaLoad[^>]*<\/script>/g,
+            '<script>function onRecaptchaLoad(){}</script>',
+        );
+        return src;
+    },
+};
+
+export default patch;
